Updating documentation, Navier-Stokes, Darcy Flow (benchmarking).
