+++ Jupyterbook/Benchmarking/NavierStokes.md
@@ -1,0 +1,17 @@
+# Navier-Stokes Benchmarks
+
+## Steady solution
+
+This benchmark is described here:
+
+
+## Periodic solution (Vortex shedding)
+
+This benchmark is described here:
+
+
+![NS Vortex shedding benchmark](Jupyterbook/Figures/Diagrams/NS_Benchmark.png)
+
+
+Standard one used in the getting started section of `Fenics`
+
