+++ Jupyterbook/Benchmarking/Stokes.md
@@ -1,0 +1,10 @@
+# Stokes Flow
+
+# Analytic benchmarking
+
+# Symmetry testing
+
+(e.g. mode coupling etc where we know which modes should be excluded)
+
+# Free surface models 
+
