+++ Jupyterbook/Figures/Diagrams/README.md
@@ -1,1 +1,3 @@
 <!-- Diagrams files - such as inkscape drawing originals / illustrator files and all that -->
+
+Note: please convert files to inkscape-compatible format (eventually) !
