+++ underworld3/meshes.pyx
@@ -155,8 +155,10 @@ class MeshFromGmshFile(MeshClass):
         # options["dm_plex_separate_marker"] = None # this is never used and flags errors for mpirun
 
         self.dm =  PETSc.DMPlex().createFromFile(filename)
-        self.meshio = meshio.read(filename)
 
+        # if the gmsh file contains periodic boundaries, we also are supposed to call this: DMLocalizeCoordinates()
+
+        self.meshio = meshio.read(filename)
 
         part = self.dm.getPartitioner()
         part.setFromOptions()
