+++ underworld3/swarm.pyx
@@ -216,6 +216,47 @@ class SwarmVariable(_api_tools.Stateful):
         return self._meshVar.fn
 
 
+    @timing.routine_timer_decorator
+    def save(self, filename : str,
+                   name     : Optional[str] = None,
+                   index    : Optional[int] = None):
+        """
+        Append variable data to the specified mesh 
+        checkpoint file. The file must already exist.
+
+        For swarm data, we currently save the proxy 
+        variable (so this will fail if the variable has
+        no proxy value). This allows some form of 
+        reconstruction of the information on a swarm
+        even if it is not an exact mapping. 
+
+        This is not ideal for discontinuous fields.
+
+        Parameters
+        ----------
+        filename :
+            The filename of the mesh checkpoint file. It
+            must already exist.
+        name :
+            Textual name for dataset. In particular, this
+            will be used for XDMF generation. If not 
+            provided, the variable name will be used. 
+        index :
+            Not currently supported. An optional index which 
+            might correspond to the timestep (for example).
+        """
+
+        # if not proxied, nothing to do. return.
+        if not self._meshVar:
+            print("No proxy mesh variable that can be saved")
+            return
+
+        self._meshVar.save(filename, name, index)
+
+        return
+
+
+
 #@typechecked
 class Swarm(_api_tools.Stateful):
 
@@ -533,3 +574,8 @@ class Swarm(_api_tools.Stateful):
                 self.data[...] = new_coords
 
         return
+
+
+
+
+
